requirements.txt
base58
python3-indy==1.8.3-dev-1109
jsonschema>=2.6.0
